Split authRouter into separate functions under words to be autoloaded
